btc: migrate bdk wallet functions to btcwallet.rs
