+++ src/server.rs
@@ -1,26 +1,21 @@
 use crate::config::ConfigOpts;
-use bdk::bitcoin::Address;
-use bdk::blockchain::{log_progress, AnyBlockchain};
-use bdk::electrum_client::{Client, ElectrumApi, ListUnspentRes};
-use bdk::sled::Tree;
-use bdk::wallet::AddressIndex::LastUnused;
-use bdk::Wallet;
 use simple_server::{Method, Server, StatusCode};
-use std::str::{from_utf8, FromStr};
+use std::str::{from_utf8};
 use std::sync::{Arc, Mutex};
 
+use crate::btcwallet::BTCWallet;
 use crate::html;
 use crate::html::not_found;
 use std::io;
 
 /// Returns a generic simple_server::Error, used to catch errors to prevent tearing
 /// down the server with a simple request, should be removed in favor of specific errors
-fn gen_err() -> simple_server::Error {
+pub fn gen_err() -> simple_server::Error {
     simple_server::Error::Io(io::Error::new(io::ErrorKind::Other, "oh no!"))
 }
 
-pub fn create_server(conf: ConfigOpts, wallet: Wallet<AnyBlockchain, Tree>) -> Server {
-    let wallet_mutex = Arc::new(Mutex::new(wallet));
+pub fn create_server(conf: ConfigOpts, btcwallet: BTCWallet) -> Server {
+    let btcwallet_mutex = Arc::new(Mutex::new(btcwallet));
 
     Server::new(move |request, mut response| {
         debug!("Request: {} {}", request.method(), request.uri());
@@ -33,12 +28,12 @@ pub fn create_server(conf: ConfigOpts, wallet: Wallet<AnyBlockchain, Tree>) -> S
             debug!("{}: {}", key, value.to_str().unwrap_or("can't map to str"));
         }
 
-        // unlock wallet mutex for this request
-        let wallet = wallet_mutex.lock().map_err(|_| gen_err())?;
+        // unlock btcwallet mutex for this request
+        let btcwallet = btcwallet_mutex.lock().map_err(|_| gen_err())?;
 
         match (request.method(), request.uri().path()) {
             (&Method::GET, "/bitcoin/api/last_unused_qr.bmp") => {
-                let address = last_unused_address(&*wallet);
+                let address = btcwallet.last_unused_address();
                 match address {
                     Ok(addr) => {
                         info!("last unused addr {}", addr.to_string());
@@ -50,7 +45,7 @@ pub fn create_server(conf: ConfigOpts, wallet: Wallet<AnyBlockchain, Tree>) -> S
                 }
             }
             (&Method::GET, "/bitcoin/api/last_unused") => {
-                let address = last_unused_address(&*wallet);
+                let address = btcwallet.last_unused_address();
                 match address {
                     Ok(a) => {
                         info!("last unused addr {}", a.to_string());
@@ -69,9 +64,7 @@ pub fn create_server(conf: ConfigOpts, wallet: Wallet<AnyBlockchain, Tree>) -> S
                 let addr = query.next().ok_or_else(|| gen_err())?;
                 let height = query.next().ok_or_else(|| gen_err())?;
                 let h: usize = height.parse::<usize>().map_err(|_| gen_err())?;
-
-                let client = Client::new(&conf.electrum_opts.electrum).map_err(|_| gen_err())?;
-                let list = check_address(&client, &addr, Option::from(h));
+                let list = btcwallet.check_address(&addr, Option::from(h));
                 match list {
                     Ok(list) => {
                         debug!("addr {} height {}", addr, h);
@@ -93,7 +86,7 @@ pub fn create_server(conf: ConfigOpts, wallet: Wallet<AnyBlockchain, Tree>) -> S
                     Some(address) => address,
                     None => return Ok(response.body(not_found().as_bytes().to_vec())?),
                 };
-                match is_my_address(&*wallet, address) {
+                match btcwallet.is_my_address(address) {
                     Ok(mine) => {
                         if !mine {
                             return Ok(response.body(
@@ -107,7 +100,7 @@ pub fn create_server(conf: ConfigOpts, wallet: Wallet<AnyBlockchain, Tree>) -> S
                 };
                 match html(
                     &conf.network.to_string(),
-                    &conf.electrum_opts.electrum,
+                    &btcwallet,
                     address,
                 ) {
                     Ok(txt) => Ok(response.body(txt.as_bytes().to_vec())?),
@@ -115,7 +108,7 @@ pub fn create_server(conf: ConfigOpts, wallet: Wallet<AnyBlockchain, Tree>) -> S
                 }
             }
             (&Method::GET, "/bitcoin") => {
-                let address = last_unused_address(&*wallet).map_err(|_| gen_err())?;
+                let address = btcwallet.last_unused_address().map_err(|_| gen_err())?;
                 let link = format!("/bitcoin/?{}", address);
                 let redirect = html::redirect(link.as_str());
                 match redirect {
@@ -139,49 +132,8 @@ pub fn create_server(conf: ConfigOpts, wallet: Wallet<AnyBlockchain, Tree>) -> S
     })
 }
 
-fn last_unused_address(wallet: &Wallet<AnyBlockchain, Tree>) -> Result<Address, bdk::Error> {
-    wallet.sync(log_progress(), None)?;
-    wallet.get_address(LastUnused)
-}
-
-fn is_my_address(
-    wallet: &Wallet<AnyBlockchain, Tree>,
-    addr: &str,
-) -> Result<bool, simple_server::Error> {
-    let script = Address::from_str(addr)
-        .map_err(|_| gen_err())?
-        .script_pubkey();
-    if wallet.is_mine(&script).map_err(|_| gen_err())? {
-        return Ok(true);
-    }
-    wallet.sync(log_progress(), None).map_err(|_| gen_err())?;
-    wallet.is_mine(&script).map_err(|_| gen_err())
-}
-
-fn check_address(
-    client: &Client,
-    addr: &str,
-    from_height: Option<usize>,
-) -> Result<Vec<ListUnspentRes>, simple_server::Error> {
-    let monitor_script = Address::from_str(addr)
-        .map_err(|_| gen_err())?
-        .script_pubkey();
-
-    let unspents = client
-        .script_list_unspent(&monitor_script)
-        .map_err(|_| gen_err())?;
-
-    let array = unspents
-        .into_iter()
-        .filter(|x| x.height >= from_height.unwrap_or(0))
-        .collect();
-
-    Ok(array)
-}
-
-fn html(network: &str, electrum: &str, address: &str) -> Result<String, simple_server::Error> {
-    let client = Client::new(electrum).map_err(|_| gen_err())?;
-    let list = check_address(&client, &address, Option::from(0)).map_err(|_| gen_err())?;
+fn html(network: &str, btcwallet: &BTCWallet, address: &str) -> Result<String, simple_server::Error> {
+    let list = btcwallet.check_address(&address, Option::from(0)).map_err(|_| gen_err())?;
 
     let status = match list.last() {
         None => "No tx found yet".to_string(),
